+++ .bunext/react-ssr/hydrate.ts
@@ -1,6 +1,6 @@
 "use client";
-import { hydrate } from "@bunpmjs/bunext/bun-react-ssr/hydrate";
+import { hydrate } from "@bunpmjs/bunext/internal/hydrate";
 import { Shell } from "./shell";
-import "./global";
+import "@bunpmjs/bunext/internal/globals";
 
 await hydrate(Shell as any);
