--- bun-react-ssr/router/components/Link.tsx
@@ -1,17 +1,0 @@
-"use client";
-import type { JSXElementConstructor } from "react";
-import { useLink } from "..";
-
-export function Link<
-  T extends keyof JSX.IntrinsicElements | JSXElementConstructor<any> = "a"
->({
-  as: Component = "a" as never,
-  href,
-  ...props
-}: { as?: T; href: string } & React.ComponentPropsWithoutRef<T>) {
-  // @ts-ignore
-  if (typeof window === "undefined") return <Component {...props} />;
-  const handleClick = useLink(href);
-  // @ts-ignore
-  return <Component {...props} onClick={handleClick} />;
-}
