--- bun-react-ssr/router/hooks/useLink.tsx
@@ -1,12 +1,0 @@
-"use client";
-import { useCallback } from "react";
-import { navigate } from "..";
-export function useLink(href: string) {
-  return useCallback(
-    (e: any) => {
-      e.preventDefault?.();
-      navigate(href);
-    },
-    [href]
-  );
-}
