+++ features/router.ts
@@ -1,3 +1,5 @@
+// only use this module in a server context
+
 const isServer = typeof window == "undefined";
 const publicThrow = () => {
   throw new Error("you cannot call revalidate in a client context");
@@ -14,17 +16,18 @@ const findRouteOrThrow = (path: string) => {
   if (!matched) throw noRouteThrow(path);
   return matched;
 };
+const noBuilderThrow = () => new Error("Builder could't be loaded");
+
+let builder = (await import("../internal/build.ts")).builder;
 
 export async function revalidate(path: string) {
   if (!isServer) publicThrow();
-  const serverModule = await import("../internal/makeBuild");
+  if (!builder) throw noBuilderThrow();
+
   const route = findRouteOrThrow(path);
-  const index = globalThis.ssrElement.findIndex(
-    (p) => p.path === route.filePath
-  );
-  if (index == -1) return;
-  globalThis.ssrElement.splice(index, 1);
-  await serverModule.makeBuild();
+  if (builder.findPathIndex(route.filePath) == -1) return;
+  builder.resetPath(route.filePath);
+  await builder.makeBuild();
 }
 /**
  *
@@ -34,9 +37,11 @@ export async function revalidate(path: string) {
 
 export function revalidateEvery(path: string, seconde: number) {
   if (!isServer) return;
-  const _revalidate = globalThis.revalidates.find((r) => r.path === path);
+  if (!builder) throw noBuilderThrow();
+
+  const _revalidate = builder.revalidates.find((r: any) => r.path === path);
   if (!_revalidate) {
-    globalThis.revalidates.push({
+    builder.revalidates.push({
       path: path,
       time: seconde * 1000,
     });
