+++ internal/globals.ts
@@ -1,3 +1,10 @@
+import type { _Head } from "@bunpmjs/bunext/componants/head";
+
+declare global {
+  var head: { [key: string]: _Head };
+  var __NODE_ENV__: "development" | "production";
+}
+
 export const paths = {
   bunextDirName: ".bunext",
   bunextModulePath: "node_modules/@bunpmjs/bunext",
@@ -14,3 +21,5 @@ export const exitCodes = {
   build: 102,
   runtime: 101,
 } as const;
+
+globalThis.__NODE_ENV__ ??= "development";
