--- internal/index.ts
@@ -1,3 +1,0 @@
-export { webToken } from "@bunpmjs/json-webtoken";
-
-export const isServer = typeof window === "undefined" ? true : false;
