+++ internal/types.ts
@@ -1,5 +1,32 @@
-import type { _GlobalData } from "../bun-react-ssr/types";
 import type { _Head } from "../componants/head";
+
+export interface ServerSideProps {
+  props?: any;
+  redirect?: string;
+}
+
+export type _DisplayMode = {
+  nextjs?: {
+    layout: string;
+  };
+  none?: "none";
+};
+export type _SsrMode = "nextjs" | "none";
+
+export const URLpaths = {
+  serverAction: "/ServerActionGetter" as const,
+};
+
+export type _GlobalData = {
+  __PAGES_DIR__: string;
+  __INITIAL_ROUTE__: string;
+  __ROUTES__: Record<string, string>;
+  __SERVERSIDE_PROPS__?: any;
+  __DISPLAY_MODE__: keyof _DisplayMode;
+  __LAYOUT_NAME__: string;
+  __LAYOUT_ROUTE__: string[];
+  __DEV_MODE__: boolean;
+};
 
 export type _globalThis = _GlobalData & {
   __HEAD_DATA__: Record<string, _Head>;
@@ -24,3 +51,12 @@ export type ServerConfig = {
 export type OnRequestType = (
   request: Request
 ) => Response | void | Promise<Response | void>;
+
+export type ssrElement = {
+  path: string;
+  elements: Array<{
+    tag: string;
+    reactElement: string;
+    htmlElement: string;
+  }>;
+};
